changed the size of the pictures to medium and moved the get helper to the client side
