+++ client/helpers/index.js
@@ -1,3 +1,4 @@
+const axios = require('axios');
 module.exports.getLocation = () => {
   // if (navigator.geolocation) {
   return navigator.geolocation.getCurrentPosition(position => (console.log({
@@ -7,4 +8,19 @@ module.exports.getLocation = () => {
   // } else {
   //   console.log('geolocation is not supported');
   // }
+};
+
+module.exports.getPic = async (pic) => {
+  try {
+    const options = {
+      method: 'get',
+      url: `https://i.imgur.com/${pic}m.jpg`,
+    };
+    const response = await axios(options);
+    const { picture } = response.data;
+    return picture;
+  } catch (err) {
+    console.log(err);
+    return err;
+  }
 };
