+++ server/helpers/index.js
@@ -60,25 +60,10 @@ const sanitizeUser = (user) => {
   return sanitizedUser;
 };
 
-const getPic = async (pic) => {
-  try {
-    const options = {
-      method: 'get',
-      url: `https://i.imgur.com/${pic}.jpg`,
-    };
-    const response = await axios(options);
-    const { picture } = response.data;
-    return picture;
-  } catch (err) {
-    console.log(err);
-    return err;
-  }
-};
 
 module.exports = {
   fetchRestaurants,
   fetchSpot,
   selectMatch,
   sanitizeUser,
-  getPic,
 };
